Move konsist test to their own module.
