+++ settings.gradle.kts
@@ -52,6 +52,7 @@ enableFeaturePreview("TYPESAFE_PROJECT_ACCESSORS")
 rootProject.name = "ElementX"
 include(":app")
 include(":appnav")
+include(":tests:konsist")
 include(":tests:uitests")
 include(":tests:testutils")
 include(":anvilannotations")
