+++ tests/konsist/src/test/kotlin/io/element/android/tests/konsist/KonsistConfigTest.kt
@@ -1,0 +1,45 @@
+/*
+ * Copyright (c) 2023 New Vector Ltd
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.element.android.tests.konsist
+
+import com.google.common.truth.Truth.assertThat
+import com.lemonappdev.konsist.api.Konsist
+import org.junit.Test
+
+class KonsistConfigTest {
+    @Test
+    fun `assert that Konsist detect all the project classes`() {
+        assertThat(
+            Konsist
+                .scopeFromProject()
+                .classes()
+                .size
+        )
+            .isGreaterThan(1_000)
+    }
+
+    @Test
+    fun `assert that Konsist detect all the test classes`() {
+        assertThat(
+            Konsist
+                .scopeFromTest()
+                .classes()
+                .size
+        )
+            .isGreaterThan(100)
+    }
+}
